+++ Libplanet.Net/Messages/NetMQMessageCodec.cs
@@ -48,6 +48,16 @@ namespace Libplanet.Net.Messages
             Peer peer,
             DateTimeOffset timestamp)
         {
+            if (!privateKey.PublicKey.Equals(peer.PublicKey))
+            {
+                throw new InvalidCredentialException(
+                    $"An invalid private key was provided: " +
+                    $"the provided private key's expected public key is {peer.PublicKey} " +
+                    $"but its actual public key is {privateKey.PublicKey}.",
+                    peer.PublicKey,
+                    privateKey.PublicKey);
+            }
+
             var netMqMessage = new NetMQMessage();
 
             // Write body (by concrete class)